QJ combo lambda (#7)
* Combine QJ Lambdas into a single Lambda in preparation for containerization
